pyproject.toml
[project]
name = "academic-tracker-mcp"
version = "0.1.0"
description = "Servidor MCP que acessa meus dados acadêmicos no Sistema Unificado de Administração Pública (SUAP)."
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "fastmcp>=3.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
]
